report
git-svn-id: https://svn.code.sf.net/p/osirix/code/osirix@4291 4e9e6e3d-4551-47eb-9dbe-a6127f9400a1

+++ DCM Framework/Networking/DCMCStoreRequest.m
@@ -95,7 +95,7 @@ htt://www.pixelmed.com
 	return self;
 }
 
-- (int)priority{
+- (unsigned short)priority{
 	return priority;
 }
